requirements.txt
streamlit
pandas
numpy
reportlab
weasyprint
requests
transformers
torch
pydantic
huggingface_hub
python-dotenv
apify_client
langchain_groq
mcp_use